Changed function to remove duplicate code
Previously used two seperate functions to handle assigning values to the
tempChannelData object for online and offline status. This resulted in
almost identical assignments, differing only in the key paths of the
value of the json parameter. I rectified this by adding a boolean
parameter to the setChannelData() function that represents online status,
and then presets the relative key path for the json parameter.